+++ doc/changelog.md
@@ -15,7 +15,7 @@ Syntax: `- short text describing the change _(Your Name)_`
 - _()_
 - _()_
 - _()_
-- _()_
+- Add tests for hooks in frontend/layers _(Lukas Anton Lakits)_
 - _()_
 - _()_
 - _()_